Implementation of basic crawler organisation.
`crawler.Crawler` is defined, where a crawler has to implement
a method that forwards documents found by the crawler to a channel.

A helper function that launches a list of crawlers concurrently,
merges their outputs into one channel, and forwards errors is
also implemented.

Finally, a test that verifies the correctness and concurrency of the
helper method is provided.